backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
scikit-learn==1.3.2
tensorflow==2.15.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
PyJWT==2.8.0
